Change the routing for the display of the cameras
This commit corrects the display of camera feeds from their respective
junction. This means that the grid of cameras displayed that was
displayed in the homepage is now displayed inside the junctions path.
The data displayed inside however is not accurate and displays the same
feed regardless of what junction is selected.