Get dynamic data from the backend for cameras
This commit fetches data for cameras for junctions from the backend
instead of having to rely on dummy data. This means that each junction
page now has it's own relevant data.

The UI also displays the active selection as to where the user is in
both the top navigation bar and side navigation bar. This makes sure the
user is not lost when navigating through the application.