Change the directory pattern for ui and lib
This commit changes the directory pattern for the ui and the lib by
moving it outside the app directory (used for routing purposes). This
change was made because of the clutter happening inside the said app
directory as the code inside these two folders were unrelated to the app
directory.