Add table to the cases page
The major change made in this commit is adding a material-react-table
with dummy data in the cases page. The table is not final yet.

Another change made was having a well type defined props for the
expanded cards than having the type of any. This made sure the component
got the props that are safely typed instead of passing any type of
values.